fix(constants): Move constants to src and fix paths
