+++ src/components/Cocktails.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { useGSAP } from "@gsap/react"
 import gsap from "gsap"
-import { cocktailLists, mockTailLists } from "../../constants"
+import { cocktailLists, mockTailLists } from "../constants/index.js"
 import CocktailLeftLeaf from "../assets/images/cocktail-left-leaf.png"
 import CocktailRightLeaf from "../assets/images/cocktail-right-leaf.png"
