+++ src/components/Navbar.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { navLinks } from "../../constants"
+import { navLinks } from "../constants/index.js"
 import { useGSAP } from "@gsap/react"
 import { gsap } from "gsap"
 import Logo from "../assets/images/logo.png"
